fix : 병합 이후 SecurityConfig가 동시에 존재하던 문제 수정 userInfo측 config 패키지에서 별도로 만들어던 SecurityConfig 관련 내용을 이관. 이전 병합 커밋에서 PATCH wikiReview 요청이 추가구현되었으나 커밋메시지에 작성되지 않아 여기에 작성합니다.
