+++ lpick-backend/src/main/java/com/notfound/lpickbackend/security/config/SecurityConfig.java
@@ -7,13 +7,20 @@ import com.notfound.lpickbackend.security.util.JwtUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.access.expression.method.DefaultMethodSecurityExpressionHandler;
+import org.springframework.security.access.expression.method.MethodSecurityExpressionHandler;
+import org.springframework.security.access.hierarchicalroles.RoleHierarchy;
+import org.springframework.security.access.hierarchicalroles.RoleHierarchyImpl;
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 
 @Configuration
 @EnableWebSecurity
+@EnableMethodSecurity(prePostEnabled = true)
 @RequiredArgsConstructor
 public class SecurityConfig {
 
@@ -38,6 +45,34 @@ public class SecurityConfig {
         http.addFilterBefore(new JwtFilter(jwtUtil), UsernamePasswordAuthenticationFilter.class);
 
         return http.build();
+    }
+
+    // 1) Tier만을 위한 RoleHierarchy 빈
+    @Bean
+    public RoleHierarchy roleHierarchy() {
+        RoleHierarchyImpl h = new RoleHierarchyImpl();
+        h.setHierarchy(
+                // hasRole로 비교할 Tier 계층구조
+                "ROLE_TIER_ADMIN > ROLE_TIER_EXPERT\n" +
+                        "ROLE_TIER_EXPERT > ROLE_TIER_DIAMOND\n" +
+                        "ROLE_TIER_CHALLENGER > ROLE_TIER_DIAMOND\n" +
+                        "ROLE_TIER_DIAMOND > ROLE_TIER_GOLD\n" +
+                        "ROLE_TIER_GOLD > ROLE_TIER_SILVER\n" +
+                        "ROLE_TIER_SILVER > ROLE_TIER_BRONZE" +
+
+                        // hasAuthority로 비교할 Auth 계층구조(ADMIN이 모든 기능 가능하게 하기위함.
+                        "AUTH_ADMIN > AUTH_MANAGER\n" +
+                        "AUTH_ADMIN > AUTH_MEDIATOR"
+        );
+        return h;
+    }
+
+    // 2) Method Security Expression Handler에 위 계층 주입
+    @Bean
+    public MethodSecurityExpressionHandler methodSecurityExpressionHandler(RoleHierarchy hierarchy) {
+        DefaultMethodSecurityExpressionHandler h = new DefaultMethodSecurityExpressionHandler();
+        h.setRoleHierarchy(hierarchy);
+        return h;
     }
 
 }
